Update fsCognitoPostConfirmation-cloudformation-template.json
Fix: Lambda function runtime updated from nodejs14.x to nodejs20.x!

The runtime of the Lambda function in the CloudFormation template created by Amplify CLI is incorrectly set to nodejs14.x. As a result of the error given during the build phase of the newly created application in AWS Amplify, the runtime was updated to nodejs20.x.